requirements.txt
Django>=5.2.1
openai>=1.14.0
gunicorn>=21.2
python-dotenv>=1.0
